+++ src/commons/profile/ProfileContainer.ts
@@ -2,7 +2,7 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 
 // TODO: Import from commons
-import { fetchAssessmentOverviews } from 'src/commons/actions/SessionActions';
+import { fetchAssessmentOverviews } from 'src/commons/application/actions/SessionActions';
 import { IState } from 'src/reducers/states';
 import Profile, { IProfileDispatchProps, IProfileStateProps } from './ProfileComponent';
