+++ src/pages/academy/AcademyContainer.ts
@@ -3,7 +3,7 @@ import { RouteComponentProps, withRouter } from 'react-router';
 import { bindActionCreators } from 'redux';
 
 // TODO: Import from commons
-import { fetchNotifications } from 'src/commons/actions/SessionActions';
+import { fetchNotifications } from 'src/commons/application/actions/SessionActions';
 // TODO: Import from commons
 import { IState } from 'src/reducers/states';
 import Academy, {
