+++ src/pages/academy/dashboard/DashboardContainer.ts
@@ -2,7 +2,7 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 
 // TODO: Import from commons
-import { fetchGradingOverviews } from 'src/commons/actions/SessionActions';
+import { fetchGradingOverviews } from 'src/commons/application/actions/SessionActions';
 import { fetchGroupOverviews } from 'src/features/dashboard/DashboardActions';
 import { IState } from 'src/reducers/states';
 import Dashboard, { IDashboardDispatchProps, IDashboardStateProps } from './DashboardComponent';
