+++ src/pages/academy/grading/subcomponents/GradingEditorContainer.ts
@@ -1,7 +1,7 @@
 import { connect, MapDispatchToProps } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 
-import { submitGrading, submitGradingAndContinue } from 'src/commons/actions/SessionActions';
+import { submitGrading, submitGradingAndContinue } from 'src/commons/application/actions/SessionActions';
 import GradingEditor, { DispatchProps } from './GradingEditorComponent';
 
 const mapDispatchToProps: MapDispatchToProps<DispatchProps, {}> = (dispatch: Dispatch) =>
