+++ src/pages/academy/sourcereel/SourcereelContainer.ts
@@ -20,7 +20,7 @@ import {
   beginInterruptExecution,
   debuggerReset,
   debuggerResume
-} from 'src/commons/actions/InterpreterActions';
+} from 'src/commons/application/actions/InterpreterActions';
 
 import {
   browseReplHistoryDown,
