+++ src/pages/login/LoginContainer.ts
@@ -2,7 +2,7 @@ import { connect, MapDispatchToProps, MapStateToProps } from 'react-redux';
 import { bindActionCreators, Dispatch } from 'redux';
 
 // TODO: Refactor
-import { fetchAuth, login } from 'src/commons/actions/SessionActions';
+import { fetchAuth, login } from 'src/commons/application/actions/SessionActions';
 import Login, { DispatchProps, OwnProps } from './LoginComponent';
 
 const mapStateToProps: MapStateToProps<{}, OwnProps, {}> = (_, ownProps) => ownProps;
