+++ src/pages/playground/PlaygroundContainer.ts
@@ -24,7 +24,7 @@ import {
   beginInterruptExecution,
   debuggerReset,
   debuggerResume
-} from 'src/commons/actions/InterpreterActions';
+} from 'src/commons/application/actions/InterpreterActions';
 
 import {
   browseReplHistoryDown,
